+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -5,6 +5,7 @@
 -   Force `<SearchListItem>` form elements id to be unique. #6871
 -   Add `controlId` and `name` props to `<SearchListItem>`. #6871
 -   Minor styling tweaks and fixes to `<SearchListcontrol>`. #6871
+-   Fix `autocompleter` for custom Search in `FilterPicker` #6880
 
 # 6.1.2